Finish create note
